+++ tests/test_forecast_application.py
@@ -48,6 +48,7 @@ class ForecastProductionTests(unittest.TestCase):
             "dataclasses",
             "datetime",
             "hashlib",
+            "heapq",
             "math",
             "statistics",
             "zoneinfo",